Basic.has(): move handling of AssocOps to AssocOp
Instead of hardcoding the matching function inside Basic,
pattern._has_matcher() is now used, thus removing a dependency
of Basic on its subclasses.
