+++ sympy/core/basic.py
@@ -1056,44 +1056,16 @@ class Basic(object):
 
     def _has(self, pattern):
         """Helper for .has()"""
-        def _ncsplit(expr):
-            cpart, ncpart = [], []
-            for arg in expr.args:
-                if arg.is_commutative:
-                    cpart.append(arg)
-                else:
-                    ncpart.append(arg)
-            return set(cpart), ncpart
-
-        def assoc_matcher(pattern):
-            c, nc = _ncsplit(pattern)
-            def is_in(expr):
-                if expr == pattern:
-                    return True
-                elif not isinstance(expr, Basic):
-                    return False
-                elif expr.is_Add or expr.is_Mul:
-                    _c, _nc = _ncsplit(expr)
-                    if (c & _c) == c:
-                        if not nc:
-                            return True
-                        elif len(nc) <= len(_nc):
-                            for i in xrange(len(_nc) - len(nc)):
-                                if _nc[i:i+len(nc)] == nc:
-                                    return True
-                return False
-            return is_in
-
         pattern = sympify(pattern)
-
         if isinstance(pattern, BasicType):
-            match = lambda expr: (isinstance(expr, pattern))
-        elif pattern.is_Add or pattern.is_Mul:
-            match = assoc_matcher(pattern)
+            return any(isinstance(arg, pattern) for arg in preorder_traversal(self))
         else:
-            match = lambda expr: expr == pattern
+            match = pattern._has_matcher()
+            return any(match(arg) for arg in preorder_traversal(self))
 
-        return any(match(arg) for arg in preorder_traversal(self))
+    def _has_matcher(self):
+        """Helper for .has()"""
+        return self.__eq__
 
 
     def replace(self, query, value, map=False):
